override keyboard localiser
